requirements.txt
python-chess>=1.999
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
stockfish>=3.28.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.5.0